+++ CHANGELOG.md
@@ -1,8 +1,24 @@
 Changelog
 =========
 
-v1.3.3
+v1.3.4
 ------
+* Fix NullPointerException when too many nodes are unevictable. A CacheExhaustedException
+  should be thrown instead.
+* Fix deadlock between node split and checkpoint.
+* Fix "Already in NodeMap" exception when loading fragmented nodes.
+* Add some default View and Cursor method implementations.
+
+v1.3.3.1 (2016-08-02)
+--------
+* Fix subtraction error when load encounters a split node, causing wrong value to be loaded.
+
+v1.3.3 (2016-07-30)
+------
+* Fixed transaction race condition which allowed shared locks to be prematurely released.
+* Fixed load race conditions which caused an incorrect value to be returned.
+* Fix for performing database compaction while old indexes are concurrently deleted. Some
+  pages would get lost, preventing compaction from ever working again.
 * Support temporary indexes.
 * Don't close in-use indexes during verification.
 * Redo decoder should be lenient if EOF is reached in the middle of an operation.